changed footer file name
